moved ClientNetwork to another package
